\name{caen-package}
\alias{caen-package}
\alias{caen}
\docType{package}
\title{
What the package does (short line)
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab caen\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-12\cr
License: \tab What license is it under?\cr
}
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}

