\name{realData}
\docType{data}
\alias{realData}
\title{A real dataset of gene expression RNA-seq.}
\description{
    Gene expression is compared between two mouse models, Pkd1f/f: HoxB7-cre 
    mice and Pkd2f/f: HoxB7-cre mice. Each group includes 18 samples, and 
    there are total 29996 transcripts in this dataset.
}
\usage{realData}
\format{A SummarizedExperiment Bioconductor object containing 29996 
    transcripts.}
\source{
    Woo, Y., Kim, D., Koo, N., Kim, Y., Lee, S., Ko, J., et al. (2017). 
    Profiling of mirnas and target genes related to cystogenesis in adpkd 
    mouse models. scientific reports 7, 14151. 
}
\keyword{datasets}