\name{sim_data}
\docType{data}
\alias{sim_data}
\title{A simulation dataset of RNA-seq.}
\description{
    This data set includes 100 samples and 474 genes.
    There are four classes in this data set.
}
\usage{sim_data}

\format{A data.frame containing 474 genes and 100 samples.}
\source{
    Zhou Y,Wan X,Zhang B,et al.Classifying next-generation
    sequencing data using a zero-inflated Poisson model[J]
}
\keyword{datasets}

