% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAEN.R
\name{CAEN}
\alias{CAEN}
\title{Compute the correlation coefficient of gene with category number to identify
differentially expressed genes}
\usage{
CAEN(data,y,K,gene_no_list)
}
\arguments{
\item{data}{Matrix or data.frame containing read counts and the row denotes
the gene}

\item{y}{the category for each sample}

\item{K}{the number of class}

\item{gene_no_list}{the number of differentially expressed genes you want
to select}
}
\value{
list(.) A list of computed correlation coefficient and the first some
differentially expressed genes , "r"represents correlation coefficient
 between gene and category number, "np" represents the top differential
 feature label.
}
\description{
To Compute the correlation coefficient of gene with category
number to identify  differentially expressed genes.
}
\examples{
library(SummarizedExperiment)
data(sim_data)
CAEN(data=assay(sim_data),y=y,K=4,gene_no_list=100)
}
