\name{realData}
\docType{data}
\alias{realData}
\title{A real dataset of single-cell RNA-seq.}
\description{
This data set gives 44 samples and 20182 genes.
}
\usage{orthgenes}

\format{A data.frame containing 20182 genes.}
\source{
Brawand D, Soumillon M, Necsulea A, Julien P, Csardi G, Harrigan P, et al.
The subiculum is a patchwork of discrete subregions.
}
\keyword{datasets}

